Fixed attributes that're crashing the client (Attack, Precision, Vitality, Defense - they should stay as ints). Added back GenericBlobMessage14 to GenericBlobMessage.cs, as gsdumper was crashing without it.

+++ src/Mooege/Core/GS/Players/Player.cs
@@ -316,10 +316,10 @@ namespace Mooege.Core.GS.Players
             this.Attributes[GameAttribute.Experience_Next] = LevelBorders[this.Toon.Level];
             this.Attributes[GameAttribute.Experience_Granted] = 1000;
             this.Attributes[GameAttribute.Armor_Total] = 0;
-            this.Attributes[GameAttribute.Attack] = this.InitialAttack;
-            this.Attributes[GameAttribute.Precision] = this.InitialPrecision;
-            this.Attributes[GameAttribute.Defense] = this.InitialDefense;
-            this.Attributes[GameAttribute.Vitality] = this.InitialVitality;
+            this.Attributes[GameAttribute.Attack] = (int)this.InitialAttack;
+            this.Attributes[GameAttribute.Precision] = (int)this.InitialPrecision;
+            this.Attributes[GameAttribute.Defense] = (int)this.InitialDefense;
+            this.Attributes[GameAttribute.Vitality] = (int)this.InitialVitality;
 
             //Hitpoints have to be calculated after Vitality
             this.Attributes[GameAttribute.Hitpoints_Factor_Level] = 4f;
@@ -1384,10 +1384,10 @@ namespace Mooege.Core.GS.Players
                 else { this.Attributes[GameAttribute.Experience_Next] = 0; }
 
                 // 4 main attributes are incremented according to class
-                this.Attributes[GameAttribute.Attack] += this.AttackIncrement;
-                this.Attributes[GameAttribute.Precision] += this.PrecisionIncrement;
-                this.Attributes[GameAttribute.Vitality] += this.VitalityIncrement;
-                this.Attributes[GameAttribute.Defense] += this.DefenseIncrement;
+                this.Attributes[GameAttribute.Attack] += (int)this.AttackIncrement;
+                this.Attributes[GameAttribute.Precision] += (int)this.PrecisionIncrement;
+                this.Attributes[GameAttribute.Vitality] += (int)this.VitalityIncrement;
+                this.Attributes[GameAttribute.Defense] += (int)this.DefenseIncrement;
 
                 // Hitpoints from level may actually change. This needs to be verified by someone with the beta.
                 //this.Attributes[GameAttribute.Hitpoints_Total_From_Level] = this.Attributes[GameAttribute.Level] * this.Attributes[GameAttribute.Hitpoints_Factor_Level];
